feat(backend): #125 add support adding products with a category

+++ apps/backend/src/main/java/no/ntnu/webshop/repository/CategoryJpaRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 import no.ntnu.webshop.contracts.category.CategoryDto;
 import no.ntnu.webshop.model.Category;
 
-public interface CategoryJpaRepository extends JpaRepository<Category, Long> {
+public interface CategoryJpaRepository extends JpaRepository<Category, Integer> {
 
   /**
    * Finds all categories and returns them in a list of CategoryDto objects.
